requirements.txt
asgiref==3.7.2
Django==5.0.3
gunicorn==22.0.0
django-recaptcha==4.0.0
django-paypal==2.1
packaging==24.1
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
typing-extensions==4.7.1
wheel==0.43.0
whitenoise==6.7.0
